# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API untuk manajemen turnamen esports Mobile Legends",
        "title": "EMBECK API - Turnamen Esports Management",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/rrq-dev/EMBECK",
            "email": "support@embeck.com"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/admin/tickets": {
            "get": {
                "description": "Mendapatkan daftar semua jenis tiket dengan opsi filter dan populate tournament details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Get All Ticket Types",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter berdasarkan Tournament ID",
                        "name": "tournament_id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Include tournament details (true/false)",
                        "name": "populate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List semua ticket types dengan tournament details (dengan populate=true)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.TicketWithTournament"
                            }
                        }
                    },
                    "400": {
                        "description": "Tournament ID format tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Gagal mengambil data dari database",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/admin/tickets/{id}": {
            "get": {
                "description": "Mendapatkan detail jenis tiket berdasarkan ID dengan opsi populate tournament details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Get Ticket Type By ID",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"64f123abc456def789012345\"",
                        "description": "Ticket ID (ObjectID format)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Include tournament details (true/false)",
                        "name": "populate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail ticket type dengan tournament details (dengan populate=true)",
                        "schema": {
                            "$ref": "#/definitions/model.TicketWithTournament"
                        }
                    },
                    "400": {
                        "description": "Ticket ID format tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Ticket type tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Memperbarui detail jenis tiket. Field yang tidak diisi akan tetap menggunakan nilai lama.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Update Ticket Type",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"64f123abc456def789012345\"",
                        "description": "Ticket ID (ObjectID format)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data tiket yang akan diupdate (partial update supported)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TicketRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket berhasil diupdate",
                        "schema": {
                            "$ref": "#/definitions/model.TicketResponse"
                        }
                    },
                    "400": {
                        "description": "Request data atau ID format tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Ticket type tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Menghapus jenis tiket dari database secara permanen. Operasi ini tidak dapat dibatalkan.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Delete Ticket Type",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"64f123abc456def789012345\"",
                        "description": "Ticket ID (ObjectID format)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket berhasil dihapus",
                        "schema": {
                            "$ref": "#/definitions/model.TicketResponse"
                        }
                    },
                    "400": {
                        "description": "Ticket ID format tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Ticket type tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/tickets": {
            "post": {
                "description": "Membuat jenis tiket baru untuk turnamen. Setiap tournament bisa memiliki berbagai jenis tiket dengan harga dan kapasitas yang berbeda.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Create New Ticket Type",
                "parameters": [
                    {
                        "description": "Data tiket yang akan dibuat",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TicketRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ticket berhasil dibuat",
                        "schema": {
                            "$ref": "#/definitions/model.TicketResponse"
                        }
                    },
                    "400": {
                        "description": "Request data tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Tournament tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/tickets/user/{userId}": {
            "get": {
                "description": "Mendapatkan semua tiket yang dimiliki oleh user tertentu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Get Tickets by User ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List tiket user",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Ticket"
                            }
                        }
                    },
                    "400": {
                        "description": "ID user tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/tournaments/{id}/tickets": {
            "get": {
                "description": "Mendapatkan semua jenis tiket yang tersedia untuk tournament tertentu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Get Tickets by Tournament ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tournament ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List tiket untuk tournament",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Ticket"
                            }
                        }
                    },
                    "400": {
                        "description": "ID tournament tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Ticket": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "64f123abc456def789012345"
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-07-16T07:28:37.016Z"
                },
                "description": {
                    "type": "string",
                    "example": "Tiket Reguler - Tribun A (Standing)"
                },
                "price": {
                    "type": "integer",
                    "example": 50000
                },
                "quantity_available": {
                    "type": "integer",
                    "example": 1000
                },
                "tournament_id": {
                    "type": "string",
                    "example": "64f123abc456def789012345"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-07-16T07:28:37.016Z"
                }
            }
        },
        "model.TicketRequest": {
            "type": "object",
            "required": [
                "price",
                "quantity_available",
                "tournament_id"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Tiket Reguler - Tribun A (Standing)"
                },
                "price": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 50000
                },
                "quantity_available": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 1000
                },
                "tournament_id": {
                    "type": "string",
                    "example": "64f123abc456def789012345"
                }
            }
        },
        "model.TicketResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Ticket type created successfully"
                },
                "ticket_id": {
                    "type": "string",
                    "example": "64f123abc456def789012345"
                }
            }
        },
        "model.TicketWithTournament": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "64f123abc456def789012345"
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-07-16T07:28:37.016Z"
                },
                "description": {
                    "type": "string",
                    "example": "Tiket Reguler - Tribun A (Standing)"
                },
                "price": {
                    "type": "integer",
                    "example": 50000
                },
                "quantity_available": {
                    "type": "integer",
                    "example": 1000
                },
                "tournament": {
                    "$ref": "#/definitions/model.TournamentBasic"
                },
                "tournament_id": {
                    "type": "string",
                    "example": "64f123abc456def789012345"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-07-16T07:28:37.016Z"
                }
            }
        },
...
        "model.TournamentBasic": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "64f123abc456def789012345"
                },
                "name": {
                    "type": "string",
                    "example": "Mobile Legends Championship 2025"
                },
                "status": {
                    "type": "string",
                    "example": "upcoming"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}